Coalesce 'morsel' method
I am really not sure about the name. I also tried 'food' and 'critter',
but 'critter' is going to be used in many ways. Both the thing that the
lady swallowed, and the chain.

Interestingly, I didn't just choose two branches and made them more
similar, I did a horizontal refactoring, coalescing the whole idea of
the morsel from each branch.

This is partly because verses 1, 2, and 8 are all quite different, and
it really felt weird to add an if/else with 3 and 4 in it. There is no
'default' morsel.